locking: introduce unlocked_watchman_root type
This is used to indicate a root that is not locked.
This commit just tackles the first order dependencies/fan out from
converting the resolve_root_or_err function to populate an
unlocked_watchman_root type.

This found 1 previously undiscovered bug in the initial subscription
results: we should have been locking for the query that was being run
there, and this may account for the occasional report of getting over
notified around the time that we initiate making a subscription.