locking: clockspec_eval needs a write lock
Evaluation of a named cursor needs to update the cursor map.

Since we're angling to allow running queries without holding a write
lock, I've added a read only variant of this function.  This is used
by the `since` expression term (not the `since` generator!) and cannot
update named cursors.

That turns out to be fine because we already detect and raise an error
during expression parsing if the user is trying to use a named cursor
in a since term.