locking: query execution context holds a read lock
This swaps the root pointer in the query execution context with a
read_locked_watchman_root pointer.

Since we don't actually acquire read locks yet, and we have a write
lock, we need a supported way to safely cast a
write_locked_watchman_root to a read_locked_watchman_root.

The w_root_read_lock_from_write() inline function is responsible
for doing this.

This is safe for a couple of reasons:

1. The read and write lock holders are binary compatible with each
   other; they both simply hold a root pointer.
2. The underlying unlock function pthread_rwlock_unlock works regardless
   of the read-ness or write-ness of the lock, even though we have
   a separate read and write unlock functions.
3. We're careful to pass a pointer to the existing lock instance
   around rather than copying it around; that way don't lose track of
   the fact that we unlocked the root.