perf: capture the slow query in the query perf sample
I thought we were already doing this, but apparently not.
This captures a reference to the incoming query so that we can send it
out along with the perf sample if that query proves to be a slow query.

To accomodate some of the queries produced by buck and logged to the log
file, we need to increase the size of the stack buffer that we use to
log a line.