Add a cache for the parent path in the query context
A recent change made it O(depth-of-tree) to compute the wholename.
We already cache the wholename in the query executor, but there are
some operations where we are also interested in the name of the
parent of a file node.

This diff adds a simple cache for the most recently looked up file
parent and its full name string.