move state saving into its own thread
There are a couple of code paths that are made tricky because they may
deadlock around saving the state information.

This removes the source of the trickiness by having a dedicated thread
for saving the state; whenever we need to save the state, we signal the
thread to wakeup and it can block and take its sweet time without fear
of impacting the other threads that have more important things to do.