if libunwind is available, prefer it for backtraces
There's an intermittent test failure on travis that looks like a crash,
but the trace information is just an address and we don't have access
to the generated binary to dig in.

This diff uses libunwind when available to make a better attempt
at getting the procedure name in the generated stack trace.

In addition, in the fault handler, we now also attempt to run addr2line
on the instruction pointer addresses in the hope that it will give
us the file and line number information.

This is a bit gross, but it is much less code than pulling in a dwarf
library and decoding this stuff for ourselves.

:-/