string: add w_string smart pointer
Adds `w_string`, a smart pointer type that automatically tracks references to
`w_string_t` and that is implicitly convertible to `w_string_t*` so that
it can be passed to our existing suite of string related functions.

Added a couple of helper methods for ensuring null-termination and
returning the dirname and basename.

Cut the new symlink tracking code over to use this new type; it
simplifies the memory management of the strings themselves.