watchman: first cut at making it build with buck in fbcode
Summary:
We can't just migrate wholesale to TARGETS because it is important to
keep our existing autoconf and windows build pieces working externally.

This diff adds probe.py as a light-weight alternative to autoconf that we can
use to probe certain things about the system and generate our config.h header
file.

We're using some native BUCK rules so that we can specify a header prefix so
that the includes will continue to work as-is without requiring an overhaul in
how the other build systems work with this code base.

This builds just the watchman binary for now.  Will follow up with plumbing in the tests.

Reviewed By: simpkins

Differential Revision: D3799750

fbshipit-source-id: ea7716ce521ec8766caf12b30771f964783a65f5