watchman: run integration tests
Summary:
This isn't as simple as using the default python test runner in our
build infra, because we do some "clever" things to expand the tests to
different flavors of bser, json and so on.

In addition, this is not as simple as just running runtests.py in the
source tree because we want it to pick up the build version of bser.

For now the approach taken here is to build a par of the bits
that we can run, and then run that as a `buck_sh_test`.

Once that starts, we chdir into the source tree to discover the tests; they are not all bundled up in the par file.

* fixup racy trigger test case
* fixup node and php tests to run in this environment
* fix linkage of constructor-registered things that was undetected until the integration tests started working

Reviewed By: simpkins

Differential Revision: D3802973

fbshipit-source-id: d1260269952689143a52c3f8ef45c60c5e5caebb