watchman: adjust tests so that they can be run and collected by our internal CI
Summary:
The main thrust of this diff is to change our funky test matrix
expansion with a differently funky way to expand the test matrix.

The idea here is that we generate test classes for the various permutations
of tests rather than re-using the same class and passing in parameters.

This allows our internal CI system to keep track of the individual tests more
easily.

This is the same approach that we use in the eden tests.

Reviewed By: sid0

Differential Revision: D3835642

fbshipit-source-id: c93a3328d8113d12e8aa1c59b772b44ff79c97a3