check for an invalid relative_root during globbing
Summary: this avoids a segfault in the case that the relative_root
cannot be resolved for the first step of a glob.

To facilitate this, we now actually look at the return value from
a generator and, (this is the slightly nasty bit), propagate the
`errmsg` field from the `query` into the results object.  This is
slightly nasty because we would ideally put this directly in the results
object.

Since we're going to flip this to C++ very soon, I'd rather just make
this code throw an exception than do the plumbing now in C.