force launchd to unload plist
Summary:
in https://github.com/facebook/watchman/issues/358 we ended up
in a state where the contents of the plist were correct but launchd
still thought that it should be monitoring the older 4.6 binary.

I don't know how we ended up in such a state, as we unload before we
write out the new plist.

My working theory is that the unload failed for some unknown reason but
we carried on anyway.

Also, this nugget from the man page:

> NOTE: Due to bugs in the previous implementation and long-standing
> client expectations around those bugs, the load and unload subcommands
> will only return a non-zero exit code due to improper usage.
> Otherwise, zero is always returned.

This is a speculative diff to add the `-F` force flag in.  I have no way
to prove whether this will solve this, only that it doesn't break
spawning watchman for me today.
Closes https://github.com/facebook/watchman/pull/360

Reviewed By: sid0

Differential Revision: D3874674

Pulled By: wez

fbshipit-source-id: daa7c1581940d286d601546d1ab892c2f1dbddcc