string: add w_string_piece class
Summary:
this is influenced by folly::StringPiece, but is close to the minimal
subset that we use.

It is used by the pathCat method as a way to represent string storage from a
variety of sources and helps to avoid making intermediate copies just for the
sake of concatenation.

I also plan on using this in a follow-on diff relating to our hash maps.

Reviewed By: sid0

fbshipit-source-id: 0f6af32c2baccbe9a811fed0042dc565b786a046