shift ownership of watchman_file
Summary:
Previously we would use the latest_file index as the primary
way to track ownership of files.  This made things complicated when
it came to aging out old deleted nodes.

This diff builds on the short series that precedes it to move the ownership to
the `files` hash in the containing directory so that there is just one way to
destroy a file.

Reviewed By: farnz

Differential Revision: D3884253

fbshipit-source-id: 47649fb35e0f2d17671435231688880d01746e64