work around dumb exclusive handle latency on windows
Summary:
Something in the VS 2015 `_mktemp` function means that we
sometimes are unable to immediately open a handle on a newly generated
file name (!)

This works around the issue by retrying after a brief pause.

This is very lame and we should pull in an implementation of mkostemp
of our own to do a better job at this.

For now, this makes the integration tests pass.

Reviewed By: farnz

Differential Revision: D3921573

fbshipit-source-id: d3d1e4d0ec85734cebfafcafa7dbe95dc00cc6cd