pull in a folly::Synchronized as watchman::Synchronized
Summary:
We don't want to take a full on dependency on folly for our
open source non-eden build, but do want to benefit from the work already
done in folly.

This pulls in a slightly cut-down version of Synchronized and puts it
in the watchman namespace.

This in turn makes it easier to decompose the giant root lock into more fine grained components.

Reviewed By: farnz

Differential Revision: D3937235

fbshipit-source-id: 34e39e072e12ae2121b1210a157ead50321cd48d