use Synchronized<> to protect inotify state
Summary:
Replaces the manual pthread locking with Synchronized<>.
This uncovered a missing lock around the move_map cleanup code.

I didn't change the move maps themselves to `unordered_map` here as I wanted to
restrict the scope of this change somewhat, but it is desirable to do that as a
follow up at some point.

Reviewed By: farnz

Differential Revision: D3937336

fbshipit-source-id: cfe2a7be9291292b37b90ba7b81cd2f88953135a