use Synchronized<> to protect trigger map
Summary:
renames the trigger map from `commands` to `triggers` and cuts over to
using `Synchronized` to protect it.

There are still a couple of bits of code that acquire the root lock around
triggers that we can't eliminate yet, but this gets us a bit closer to breaking
apart the giant root lock.

Reviewed By: farnz

Differential Revision: D3937473

fbshipit-source-id: 1992287a8d55a9932bbb49dfbf343773468c0697