add InMemoryView class
Summary:
The intent is that we will locate the data needed to
keep track of the filesystem in this class.  This initial diff
moves just the tree, the recency index and the suffix index
over.

Follow-on diffs will move "methods" that are currently root-centric
to be methods of this class.

Reviewed By: farnz

Differential Revision: D3937614

fbshipit-source-id: babe4b71715e5fd8c388140d0c4224510d5fc233