add InMemoryView::markFileChanged
Summary:
This moves most of the guts of the mark_file_changed function into a method of the InMemoryView.

We can't move it wholesale because this action also needs to ask the watcher layer to cancel a watch on a file.
I'll make some changes to the watcher layer in following diffs to make that easier to deal with.

Reviewed By: farnz

Differential Revision: D3937670

fbshipit-source-id: ec66840a81128c49801da1a325f015a0b19658e4