start moving io_thread bits into their own struct
Summary:
As part of moving towards not requiring an InMemoryView, we also need
to identify the things that are used to build up that view.

This diff moves io_thread specific things into their own struct; in a follow-on
diff we will make starting the io thread a watcher specific thing.

Reviewed By: farnz

Differential Revision: D4020100

fbshipit-source-id: fe33109e4a1a32202e018c9e75caca2e8eb1c440