factor cookie sync out into its own class
Summary:
This will aid in subsequent refactoring.

One important difference here is that we no longer need to acquire a write lock
to the root; the cookie sync was abusing this to generate a serial number for a
unique cookie file name.  We now just use a local atomic integer for that
purpose instead.

As a result of that, we no longer need to record the root number in the
generated cookie name, so we're back to the original cookie format from before
adding root numbers, which is closer to those asserted in
`tests/integration/othercookies.php`.

Reviewed By: farnz

Differential Revision: D4045069

fbshipit-source-id: 84bf0589cb2a4a58f571847649897c2f8ad4f070