switch jansson hashtable -> std::unordered_map
Summary:
This is mostly straightforward, but note:

1. We're now totally unordered in the object storage (potentially a minor perf
gain, but we're in no position to measure this right now).

2. There were some tests that relied on some accidental determinism
in the way that the hashtable was ordered.  I've either added explicit
sorting or removed the test cases where no determinism is available (eg: the
bser serialization code does not support a way to request this)

Reviewed By: farnz

Differential Revision: D4045118

fbshipit-source-id: 2b9b4ecbc05d40661db50ab89f0c8fe3d5ebb477