introduce json_ref
Summary:
Adds a smart pointer holder for `json_t*` and make the
reference-owning functions return it.

Changes the internals storage of the json object and array types
to use json_ref.  The meat of the changes are in the jansson code
to accomodate this.

Outside of this, it is pretty mechanical, but pretty long:

1. Prefer to use `auto` for the return values of the json functions
2. Where we have `goto` crossing initialization and preventing the
   use of `auto`, explicitly use `json_ref`.
3. Prefer to explicitly `std::move` values into functions like
   `set_prop` and `send_and_dispose_response`

bigdiff

Reviewed By: farnz

Differential Revision: D4045125

fbshipit-source-id: 4efe225fc3c2baaa066b4648c9a30860237325cf