fix json string equality testing bug
Summary:
Just noticed that this was using strcmp which won't have the correct results
for sliced strings that have found their way into a json value.

This is unlikely to happen in practice, but it's worth fixing.  Also: we have a
couple of micro-optimizations for comparing w_string values, so this may end up
very slightly perf positive.

Reviewed By: farnz

Differential Revision: D4045252

fbshipit-source-id: 75a7b219db77e2f84bea5f45ec7b1f6f73a2094c