add Configuration helper class
Summary:
In order to move around the code used by InMemoryView, we need to
provide access to the configuration.  Rather than pass a reference to the root
(which is what we currently use to refer to the "local" configuration for a
root) into something contained in the root, the plan is to pass in a
Configuration object that holds a ref on the underlying json from the
`.watchmanconfig` file.

This diff just adds the helper class; it will be adopted by a follow-on diff.

Reviewed By: farnz

Differential Revision: D4045267

fbshipit-source-id: 7d5d0c4b24ba662644912c12df8c40cfb34eb594