implement signalThreads for inotify, kqueue
Summary:
This allows us to eliminate the pthread_kill that we currently use to
interrupt syscalls in the notify thread; that isn't portable and isn't a strong
enough semantic for the refactoring that we want to do.

fsevents and win32 already implement this feature; this is just bringing us into parity.

Reviewed By: farnz

Differential Revision: D4045463

fbshipit-source-id: 901f23d496656921dba4a959513a032cb49c12fe