ht: fix w_ht_iter_del to not skip over entries
Summary:
This was happening because if the first entry in a bucket was deleted, we'd
skip over that bucket entirely. Subtract 1 from the slot to make sure we retry
the same slot.

Included are both unit and integration tests demonstrating the issue.

I believe I've ticked all the boxes (buck, Windows), but please let me know if I missed something.
Closes https://github.com/facebook/watchman/pull/379

Reviewed By: wez

Differential Revision: D4057012

Pulled By: sid0

fbshipit-source-id: 50ad60a6b1ea3a5272101fa8ab53be21c598ea92