move io/notifythread to InMemoryView
Summary:
These two threads are closely coupled to maintaining the InMemoryView, so move them there.

There are some notable changes here:

1. The lifetime management of the View has become more complex; since a recrawl
can cause a new View to be created from a method inside a View, we cannot join
these threads.  Instead we need to keep the View alive via a shared_ptr ref and
allow the threads to self terminate.

2. We used to remove the root from the global watch list when the notifythread
exited.  This used to only happen when the root was cancelled, but now that it
is part of the view it happens on a recrawl.  Because of this, the removal code
now happens inside w_root_cancel(), and this has a little bit of fanout to the
function that is used to delete all current watches.

I took this opportunity to use std::thread to spawn these threads.

Reviewed By: farnz

Differential Revision: D4045549

fbshipit-source-id: f30fdf2c61d2bbe01f0c3159766f2fde55a70a2a