reduce dependency on write_locked roots in clockspec code
Summary:
I'm angling to remove write_locked_watchman_root in favor of
Synchronized<> access the various members of the root.

The clockspec code requires a write lock in the flavor of query execution that
uses named cursors.  It doesn't really need a write lock on the root as it
obtains a write lock on the cursor map for itself, except for the two code
paths that I've changed in this diff.

They were both incrementing the tick counter in the root with the thought that
this would avoid seeing the same results repeated over and over again.  This
turns out to be overly pessimistic for two reasons:

1. The timeGenerator code will always return items that are >= the tick used as the basis
2. Query synchronization touches a file that will bump the tick counter anyway.

I've removed the tick incrementing code from this portion.
The cursor update code just needs to be able to sample the most recent tick value from the view.

Reviewed By: farnz

Differential Revision: D4049300

fbshipit-source-id: 9bd8800eefd0c52ee91a74e550d67074945596d8