remove json_pack API in favor of std::initializer_list
Summary:
A couple of bugs in the past 6 months and during recent refactoring
have been due to mismatches between the format string and parameters passed to
json_pack.

Now that we have the json_ref type, all uses of json_pack can be replaced with
object or array construction from a list of parameters.

This diff introduces `json_object` and `json_array` functions that accept a
suitable list of parameters and return a json ref to the appropriately
constructed result, in a type-safe manner.

Reviewed By: farnz

Differential Revision: D4064975

fbshipit-source-id: 53a34fbd11598866784fb3c48703b9c73304303a