move markDirDeleted from handle_open_errno
Summary:
handle_open_errno currently reaches in to the view to mark a dir
deleted.  This is posing some problems with refactoring because we'll need to
be able to pass a write-locked ref to the view guts down into here.

Looking at this function more closely, we see that it is only called to handle
error paths in Watcher::startWatchDir implementations.  This particular
function is only called from inside crawler, in a method of InMemoryView.
Therefore is is valid to just move the markDirDeleted call up to the indirect
caller in the nullptr return code path.

Reviewed By: farnz

Differential Revision: D4073395

fbshipit-source-id: f786c6aee2fb974641e9595372cf57aeee0c4fab