watchman_user_client -> proper c++ inheritance
Summary:
This stuff is a bit fugly.  The intent was to re-use the guts
of listener.cpp with the nascent direct-to-fsevents flavor of watchman
and have a registerable factory create a watchman_user_client or a
watchman_fsevents_raw_client depending on the context of the process.

We didn't quite get to implementing the latter.

Now that we're in c++ land, we can tidy up the construction/destruction of this
stuff by making it a virtual class.

This diff "breaks" the factory portion of this functionality, but that can be
restored when we get back to the direct server stuff.

Reviewed By: farnz

Differential Revision: D4073535

fbshipit-source-id: f43f1a8d9a2dab3308ba8db1d2186ef1311e682e