art: make ValueType templatizeable
Summary:
I'm going to follow up with move construction and perfect forwarding;
this is just the basics of storing typed values.

Note that the search API returns a pointer to the stored item, which is a
semantic change from previously; when `ValueType=void*` this used to return the
value or nullptr (which was ambiguous).  It now returns the address of the
stored value, or nullptr if no item was found.

Reviewed By: farnz

Differential Revision: D4075891

fbshipit-source-id: 0a34773722d44a84c820be62c308b3aeb66c008c