art: tidy up formatting of this bit of code
Summary:
The vim modeline keeps messing with me, so remove it, and fix up the
formatting around this RECURSE_SEARCH label.

I've also removed the top-of-scope `Leaf* l` variable in favor of variables
more closely scoped to their use; this will come in handy when adding smart
pointers in here.

Reviewed By: farnz

Differential Revision: D4076563

fbshipit-source-id: c1d97b5f9f62c10a366e2c69e7dfdf99a64bc050