Synchronized<> pending collection
Summary:
Adopt Synchronized for locking.  `Synchronized<T, std::mutex>` has a
specialization that allows access to the underlying unique_lock so that we can
pass it to condition_variable.

Break up the pending collection into a base that is protected by the lock
and make PendingCollection derived from Synchronized<Base> so that we can
add a helper to wait on the condition variable.

We need to pass down a reference to the condition and the pinged predicate into
the Base to allow its ping method to work both for the locked version and also
because in one case we need to do a non-blocking (wrt. this mutex) ping when
we're signalling threads.

Reviewed By: farnz

Differential Revision: D4080776

fbshipit-source-id: ce8df950e302b24ebb45469b5f782263a13fb941