add w_string::build as alternative to w_string::printf
Summary:
This is a type-safe way to build up a string from a variety of
parameters.

It's not as powerful in terms of formatting flexibility as printf, but does
mean that we don't need to think so hard about printf format strings.

The templates make this appear more complex than it really is; there are
two sets of helper functions, one to measure how much (worst case) space
is needed, and another to emit the string value into the buffer.

Reviewed By: farnz

Differential Revision: D4102017

fbshipit-source-id: 7e75cc6d6cd6c02e334b11bc0dfb68d83ab5c4e4