re-jigger log subscriptions
Summary:
while analyzing the TSAN output for the build, it became apparent that
we have some potential lock inversion code paths around how we inject data into
the client response streams.

This diff is the first part of resolving that issue.  There will be a follow-up
diff that changes the file change subscriber and state-enter plumbing to use
PubSub too.

Previously, if client had turned up debugging logging on their subscription, we
would discover this by locking the set of clients and walking them.  This opens
up to risk of lock inversion if we are running a command that locks the client
list and then takes some other action (or even just logs something).

To resolve this, this commit introduces a PubSub helper class.  The purpose of
this class is to record a stream of json objects to be consumed by some number
of subscribers.  If there are no subscribers, nothing needs to be recorded.

If there are subscribers, an immutable Item is logged to the stream and tagged
with a new serial number.

Each Subscriber remembers the last serial number that it has observed, allowing
the `getNext` method to return the next un-observed item on each subsequent
call.

Subscribers are registered along with a callback that is intended to "wake up"
the subscriber so that it can process the newly arrived items.

This diff cuts logging over to this new PubSub scheme.  The new Log class
maintains two separate Publishers; one for debug level events and another for
error level events.

It also establishes subscriber(s) for these dependent on the global log level
flag.  These subscribers have a "notification" callback that sorts and logs any
pending items to the stderr stream, optionally aborting the process if we saw a
FATAL level log.

Each connected watchman client also establishes subscribers for the debug and
error streams if the client issues a `log-level` command.  The notify callback
for these pings the client so that it can wake up and translate the log items
into its queue of responses to send to the peer.

There are a couple of important properties:

* wlocks are needed only to amend the subscriber list or to append a new item.
* rlocks are held when a subscriber consumes items from the stream.
* The critical sections are short, non-interleaved and non-recursive.

There are some bonus side effects:

* We can avoid formatting the bulk of the log string if there are no
  subscribers for a given level.

There's a bit of fan-out from these changes:

* We now track clients via shared_ptr so that we can safely manage
  the Subscriber -> client flow that is required for the notify
  routine for the clients.  Without this we would hit a use-after-free
  each time that we tear down a client.
* To facilitate this, we need to use std::thread to spawn the client
  threads.  There's a slightly gross bit of code to capture
  pthread_self() at the start of the client thread to allow the
  pthread_kill(SIGUSR1) stuff to work.  I want to remove this
  at some point as it is not a portable way to wake up a thread.
* There's some circular log -> json -> string dependencies that mean
  that we can't use the logging functions in some lower level
  classes and helpers, so those had to be adjusted.

Reviewed By: farnz

Differential Revision: D4108737

fbshipit-source-id: e69bbfa13f42c28939855e7c19e84c0d4a5854e1