subscriptions now use PubSub
Summary:
I'm unreasonably excited about this change, as it means that we can in
theory be executing the subscription queries in parallel if there are multiple
subscriber client sessions (eg: Nuclide and hh_server).

* Change state-enter, state-leave to broadcast the state changes.
* Change the call site for process_subscriptions to instead broadcast that the root is settled.
* The listener code listens for these broadcasts on a per-client basis and then
  applies them to the appropriate client/subscription.

I had to adjust the lifetime management so that a subscription instance holds a
reference to the watched root, otherwise we can try to dispatch something from
a freed root.  This in turn meant that we had to move the
`w_cancel_subscriptions_for_root` call from the root destructor to the
`w_root_cancel` function, where the cancellation was initiated.

I plan to change triggers to use a similar mechanism in a follow on diff.

Reviewed By: farnz

Differential Revision: D4113621

fbshipit-source-id: 2a8979624cc4324b8c27b4809282b0ffb37728d3