fixup windows integration tests
Summary:
the recent change to the QueryableView class introduced paths like
'.hg/wlock' that we test as part of the defer_vcs mode in subscription
settling.

On windows, our dirName method would only consider `\\` to be the dir
separator, so this broke and the associated test started to fail.

This diff standardizes on using the `is_slash` inline to test for the dir
separator so that we check for both forward and backward slashes and avoid this
kind of problem in the future.

In addition, it makes the windows `\\` check windows platform specific so that
this doesn't impact performance on posix systems.

I've tweaked the default concurrency to reduce the flakiness and enabled
running this tests in our internal CI so that we catch this sort of problem
sooner in future.

Reviewed By: farnz

Differential Revision: D4129637

fbshipit-source-id: fadc7f5e41afef38b8e28acdd621da163506fdb4