move these file functions to be methods
Summary:
and fixup construction of watchman_file.  Since this struct
is variable size and has an embedded string at the end, we cannot
allow it to be constructed except via the provided factory method.

Delete the copy and copy-assignment functions to prevent accidentally making
copies, and tidy up destruction by adding a proper destructor and changing the
function that is called by the Deleter to "simply" perform a placement-delete
followed by releasing the storage.

Reviewed By: farnz

Differential Revision: D4140703

fbshipit-source-id: 481cdde21852aa69b01699a1ad3c58775a44d6bf