tidy up field list parsing
Summary:
Centralize this (for the most part) in the main query parsing routine, and capture the field list into the returned query object.

There are two notable exceptions to this:

1. Legacy query syntax (which has a fixed and very large list of fields)
2. Triggers (which have a specialized syntax for the stdin of the process)

Both of these cases will replace the default result of the query parse.

This is leading towards some changes in the way that we compute the query
results that will make it easier to hold on them without holding a lock on the
underlying InMemoryView.

Reviewed By: farnz

Differential Revision: D4141003

fbshipit-source-id: ea207309f94d11ac1af86b44131d6c93c3ea9ade