change query results to a json array
Summary:
Rather than build up an intermediate array holding references to the
watchman_file and some auxilliary data, change the query executor to build the
json values as the query runs.

This makes it easier in the move to eden, because we won't have to add locking
on a per result basis in this model; the json values will be immutable once the
query has run, whereas the current rule_match and file objects may be mutated
after the lock on the internal view has been released.

This is a pretty straight-forward change from the deque of results except for
one case, which is where we append the list of file names to a spawned trigger
process.  Triggers may be configured to pass a list of filenames via argv and a
user-defined list of fields that is not guaranteed to include the file name as
a json blob on stdin.  The handle that case we're mildly abusing the filename
dedup mechanism to get a list of all of the file names in the result set.
I have mixed feelings about this, but it doesn't feel excessively terrible.

Reviewed By: farnz

Differential Revision: D4142156

fbshipit-source-id: 91b36ba891e085ce775038361eb3b42198d9253d