move the Watcher totally into the InMemoryView
Summary:
The InMemoryView is now completely responsible for creating
and interacting with the Watcher class.

The WatcherRegistry now constructs and returns a View instance.  Since the
configuration for specifying which implementation to use in the
`.watchmanconfig` file uses the term "watcher", we preserve this in the name of
the WatcherRegistry.  In the near future we will add an EdenView for the eden
integration.

I can (but don't plan to do so immediately) follow up with a diff to rename the
current Watcher class to be something like `WatcherImpl` to (perhaps) improve
the clarity of the naming.

Reviewed By: farnz

Differential Revision: D4160677

fbshipit-source-id: c26c47d9ae3107c4f155b28d3a5086b16ac29490