add view::waitUntilReadyToQuery method
Summary:
This is likely only really needed for InMemoryView.

Prior to this diff there is a potential race condition where watch initiation
can return a watch response to the client before the initial crawl is complete.
This was because nothing was checking to see if the root was ready to query, it
was just coincidence that it tended to be ready due to the way that the global
root lock was acquired.  This is an issue that impacts buck from time to time,
which was worked around by running `watch-project` twice in a row(!).

This diff avoids this issue by adding a more formal mechanism for some number
of clients to synchronize with the (re)crawl state of the InMemoryView.

I've used C++11 `std::shared_future` for this.

Reviewed By: farnz

Differential Revision: D4160794

fbshipit-source-id: b1b99544a6883fe7f3ef79cf2739c899e0443656