physically remove the root lock
Summary:
We now track the root via a shared_ptr.

This diff is largely changing the following to shared_ptr<w_root_t>:

* `w_root_t*`
* `unlocked_watchman_root`
* `read_locked_watchman_root`
* `write_locked_watchman_root`

There are a couple of bootstrapping code paths that need to continue using
`w_root_t*` because they happen too early to capture a working `shared_ptr`
reference.

There are a couple of FIXME's that I'll tackle in the very next diff: there are
some slighly unsafe reads to produce clockspec strings.

Reviewed By: farnz

Differential Revision: D4160823

fbshipit-source-id: 3471af892fd8adc749a5b0c3df25397e72947848