move root number + helpers into InMemoryView
Summary:
This addresses most of the FIXMEs introduced by the prior diff by
moving the root number and tick counter into the synchronized internal view
state in the InMemoryView.

Helper methods have been added to the QueryableView to obtain the current
clock string and to obtain a pair representing the root number and tick
value.

I've introduced a couple more FIXMEs related to the clock strings that
I'm going to resolve in the next diff by adjusting the return value
from the query functions.

Reviewed By: farnz

Differential Revision: D4160844

fbshipit-source-id: 27908ee38325914f2be79f4c8c27b51df6da9b50