fix memory usage regression
Summary:
Since moving to using maps, we've been using more memory than we would
like.   This is because the caller of this function builds up the full
directory path string and then passes a slice with the baseName of it in here.
When we first register the child we will use that baseName slice, keeping alive
the longer path name as a side effect.

This diff avoids that by using the baseName slice for the lookup, but storing
the file->getName() string as the key in the case that we missed on lookup.

Reviewed By: farnz

Differential Revision: D4171999

fbshipit-source-id: a189129c95a3d707ecb227562b2dcc29b2047c8a