add port of folly/ScopeGuard.h
Summary:
This helps to resolves some of the old C-style goto cleanup,
especially now that we're in the middle of transitioning to a world with C++
exceptions.

We can't simply depend on folly at this time; hopefully the dependency and
build situation improves over the next few months and we can just do that.
In the meantime, include a port of the appropriate functionality.

Cut over the command dispatcher which was the most recently observed (D4177989)
piece of code that really needs this.  We can incrementally cut other bits of
code over as we discover them.

Reviewed By: farnz

Differential Revision: D4180389

fbshipit-source-id: 1996a6f5ac8a2e38b6a7f7c6b2f6d72b422658a5