refactor: reuse FileDescriptor, add Pipe class
Summary:
Moves the FileDescriptor class out from the kqueue watcher and expands
it a little so that we can use it to own file descriptor references in more
places.  Adds a Pipe helper class that is simply a pair of FileDescriptors with
a read and write end.

This allows us to remove a few explicit destructors and reduces the chances of
leaking descriptors.

I took the opportunity of making use of the Linux `pipe2` function, which
allows us to atomically create a pipe with the CLOEXEC bit set.

Important! the pipes that we create in this way have the non-blocking bit set.
This matches the behavior of the event object that we set up in `w_event_make`,
which is really the successor for the explicit pipe use case in watchman; we're
using pipes as objects that we can notify and wake up a `poll()`.  I think these
pipes *should* already have been set to non-blocking, so this is likely a bug
fix.  I may well eat my words, but I'm pretty confident that this is a positive
change.

Reviewed By: farnz

Differential Revision: D4181762

fbshipit-source-id: add07a515353dca9bab950097e0693c00d9453ab