c++ify watchman_event
Summary:
on Windows we used to just cast HANDLE <-> w_evt_t because they were both pointer
sized.  Now that we have a virtual base class we need to wrap up the event,
so there's a little bit more code in there.

on unix use a pipe for the API-created event object and used to make a fake
one of those to use together with the poll equivalent API.  I've teased out
the fake event object to make it clearer how it is used.

Reviewed By: farnz

Differential Revision: D4183220

fbshipit-source-id: 639cfcac46a005b004312bc9560c76f95ce799b1