add Win32Handle class
Summary:
This is similar to the FileDescriptor class but holds win32 HANDLE
objects.

In an effort to appease our coding style linters, I've opted to store the
handle value in an `intptr_t` rather than using the `HANDLE` type.  This
means that we can avoid pulling `windows.h` which is something that impacts
include ordering.

The choice of `intptr_t` mirrors the return value of the library function
`_get_osfhandle` which extracts the HANDLE value from a file descriptor.

The downside of this is that we need to cast the handle value when passing
it to the underlying win32.  I think this is a reasonable tradeoff to make.

Reviewed By: farnz

Differential Revision: D4185035

fbshipit-source-id: 79afcaa9ef7fb3d15d80a4be087c5ae145be9295