refactor w_dir_open
Summary:
This function will now throw std::system_error if it fails.
The functions that operate on the dir handle are now methods.

The windows dir reading code now targets this layer instead of trying
to implement the `opendir(3)` interface.

This makes the code around this stuff a bit nicer; the `handle_open_errno`
function is now only called from a single place to process exceptions that
might be thrown from the watcher startWatchDir method.

Reviewed By: farnz

Differential Revision: D4201434

fbshipit-source-id: fc60a55d2ccc5defbac3810949cdfff218aa7d1e